게시물 볼때 수정, 삭제 버튼 권한자만 보기

+++ list_view.php
@@ -1,15 +1,22 @@
 <?php
 # 해당 게시판 내용 보기
 
+session_start();
+
 require_once 'DB.php';
 $conn = db_connect();
+
+$user = $_SESSION['memberId'];
+$admin = "admin";
+
 $id = $_GET['id'];
-$sql = "SELECT title, content FROM board WHERE id=$id";
+$sql = "SELECT title, writer, content FROM board WHERE id=$id";
 //$board = $conn->query($sql);
 $result = mysqli_query($conn, $sql);
 $board = mysqli_fetch_array($result);
 $title = $board['title'];
 $content = nl2br($board['content']);
+$writer = $board['writer'];
 //$board = mysqli_fetch_row($stmt);
 ?>
 
@@ -19,7 +26,7 @@ $content = nl2br($board['content']);
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css" integrity="sha384-9aIt2nRpC12Uk9gS9baDl411NQApFmC26EwAOH8WgZl5MYYxFfc+NcPb1dKGj7Sk" crossorigin="anonymous">
-    <!-- <link rel="stylesheet" href="css/listViewStyle.css"> -->
+    <link rel="stylesheet" href="css/listViewStyle.css">
     <title>Document</title>
 </head>
 <body>
@@ -30,6 +37,10 @@ $content = nl2br($board['content']);
       <th scope="row" class="th-title">제목</th>
       <td><?= $title ?></td>
     </tr>
+    <tr>
+      <th scope="row" class="th-writer">작성자</th>
+      <td><?= $writer ?></td>
+    </tr>
      <tr>
       <td colspan="2" class="content"><?= $content ?></td>
     </tr>
@@ -37,8 +48,13 @@ $content = nl2br($board['content']);
 </table>
    <div class="col-auto submit submit-btn">
         <button type="button" class="btn btn-outline-secondary submit-btn__list" onclick="location.href='list.php'">목차</button>
-        <button type="button" class="btn btn-secondary submit-btn__delete" onclick="location.href='process_delete.php?id=<?= $id ?>'">삭제</button>
-        <button type="button" class="btn btn-secondary submit-btn__modify" onclick="location.href='modify.php?id=<?= $id ?>'">수정</button>
+        <?php if($user==$writer || $user==$admin) {
+          echo $user;
+          echo $writer;
+          echo $admin; ?>
+          <button type="button" class="btn btn-secondary submit-btn__delete" onclick="location.href='process_delete.php?id=<?= $id ?>'">삭제</button>
+          <button type="button" class="btn btn-secondary submit-btn__modify" onclick="location.href='modify.php?id=<?= $id ?>'">수정</button>
+        <?php } ?>
     </div> 
    </article>
